+++ src/FrontEnd/script.js
@@ -1,3 +1,3 @@
-import { router } from './router/Router.js';
+import { router } from './router/router.js';
 
 router.start();
